+++ CHANGELOG.md
@@ -10,7 +10,9 @@ This release introduces the support for Unity 6 and extends the plugin with proj
 
 ### Added
 
-- Support for Unity 6 was added
+- Support for Unity 6 was added.
+- Added project configuration checks to assist adjusting settings for AirConsole.
+- Added minimal AndroidManifest.xml upgrade logic for Unity 6
 
 ## [2.5.3] - 2024-09-27