+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Release notes follow the [keep a changelog](https://keepachangelog.com/en/1.1.0/
 
 ## [Unreleased]
 
+
+## [2.5.4] - 2024-12-10
+
 This release introduces the support for Unity 6 and extends the plugin with project configuration checks.
 
 ### Added